refactor: move sandboxed code-execution helpers into src/utils/sandbox.ts
Extract sanitizeCodeBlock, hasPlaywrightCommands and createSandboxedFunction
out of action.ts into a dedicated utils module, pushing the command-head and
shadowed-globals lists into named module-private consts (ALLOWED_COMMAND_HEADS,
PLAYWRIGHT_COMMAND_HEADS, SHADOWED_GLOBALS) used via .some(startsWith). Behavior
is identical; action.ts now imports them and no longer holds a new Function site.
